r/offset_monitor: make templated so it can be used for kafka offsets
Make offset_monitor more universal so that it can be used for different
data types.
