+++ src/v/raft/state_machine.h
@@ -140,7 +140,7 @@ private:
 
     ss::io_priority_class _io_prio;
     ss::logger& _log;
-    offset_monitor _waiters;
+    offset_monitor<model::offset> _waiters;
     model::offset _next;
     model::offset _bootstrap_last_applied;
 };
